Fixed registration

+++ include/registrator_class.hpp
@@ -25,7 +25,7 @@ public:
 
 	Registrator(std::string broker_address, service_type_t service, 
 		uint16_t reg_port, zmq::context_t *ctx);
-	uint16_t registration();
+	int32_t registration();
 	~Registrator();
 };
